requirements.txt
redis
langchain-core
langchain-redis
langchain-ollama
fastapi[standard]
python-dotenv